refactor: rename `data.ts` to `projects.ts`
