--- src/data/data.types.ts
@@ -1,14 +1,0 @@
-export interface Project {
-  id: number
-  title: string
-  description: string
-  imageUrl: string
-  website: {
-    name: string
-    url: string
-  }
-  source: {
-    name: string
-    url: string
-  }
-}
