+++ src/data/projects/projects.ts
@@ -1,4 +1,17 @@
-import { Project } from './data.types'
+export interface Project {
+  id: number
+  title: string
+  description: string
+  imageUrl: string
+  website: {
+    name: string
+    url: string
+  }
+  source: {
+    name: string
+    url: string
+  }
+}
 
 export const projects: Project[] = [
   {
